requirements.txt
whisperx>=3.8.0
openai>=1.30.0
python-dotenv>=1.0.1
gradio>=4.0.0
